src/pyproject.toml
[tool.poetry]
name = "webtronics-test"
version = "1.0.0"
description = ""
authors = ["deniko1985 <deniko1985@gmail.com>"]
readme = "README.md"
packages = [{include = "webtronics_test"}]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.10.2"
anyio = "3.6.2"
asyncio = "3.4.3"
asyncpg = "0.27.0"
bcrypt = "4.0.1"
certifi = "2022.12.7"
cffi = "1.15.1"
click = "8.1.3"
cryptography = "3.4.8"
databases = "0.7.0"
dnspython = "2.3.0"
ecdsa = "0.18.0"
fastapi = "0.94.0"
fastapi-jwt-auth = "0.5.0"
greenlet = "2.0.2"
h11 = "0.14.0"
httpcore = "0.16.3"
httpx = "0.23.3"
idna = "3.4"
jwt-pydantic = "0.0.6"
loguru = "0.6.0"
mako = "1.2.4"
markupsafe = "2.1.2"
passlib = "1.7.4"
psycopg2-binary = "2.9.5"
pyasn1 = "0.4.8"
pycparser = "2.21"
pydantic = "1.10.6"
pyjwt = "1.7.1"
python-dotenv = "1.0.0"
python-jose = "3.3.0"
python-multipart = "0.0.6"
rapidfuzz = "2.15.0"
redis = "4.6.0"
requests = "2.28.2"
rfc3986 = "1.5.0"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "1.4.46"
starlette = "0.26.0.post1"
typing-extensions = "4.5.0"
urllib3 = "1.26.15"
uvicorn = "0.21.0"
email-validator = "^2.0.0.post2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
